youview/yvrig-control: Add a yvrig-control udev rules extension
Add a system extension that adds YouView udev rules to the operating system
filesystem. The rules mount USB video capture devices to a deterministically
named video device on the system. The extension does not contain any kernel
specific files so it does not need to be rebuilt for different base kernels or
Talos versions.

AUTO-15586